+++ crates/remote/src/remote_service.rs
@@ -176,7 +176,7 @@ impl RemoteService {
         // bump the total body size larger than the actual limit. Is there a better
         // way to handle this? Probably, but for now, just reduce the size by 1%,
         // which is about 42k bytes.
-        max - (max as f64 * 0.01) as i64
+        max - (max as f64 * 0.0125) as i64
     }
 
     #[instrument(skip(self, state))]
@@ -192,9 +192,9 @@ impl RemoteService {
     }
 
     #[instrument(skip(self, state))]
-    pub async fn save_action(&self, state: &mut ActionState<'_>) -> miette::Result<()> {
+    pub async fn save_action(&self, state: &mut ActionState<'_>) -> miette::Result<bool> {
         if !self.cache_enabled {
-            return Ok(());
+            return Ok(false);
         }
 
         let missing = self
@@ -216,19 +216,30 @@ impl RemoteService {
                 .await?;
         }
 
-        Ok(())
+        Ok(true)
     }
 
     #[instrument(skip(self, state))]
-    pub async fn save_action_result(&self, state: &mut ActionState<'_>) -> miette::Result<()> {
+    pub async fn save_action_result(&self, state: &mut ActionState<'_>) -> miette::Result<bool> {
         if !self.cache_enabled {
-            return Ok(());
+            return Ok(false);
         }
 
         let Some((mut result, blobs)) = state.extract_for_upload() else {
-            return Ok(());
+            return Ok(false);
         };
 
+        // If no blobs, upload in the current thread
+        if blobs.is_empty() {
+            let value = self
+                .client
+                .update_action_result(&state.digest, result)
+                .await?;
+
+            return Ok(value.is_some());
+        }
+
+        // Otherwise upload in the background
         let client = Arc::clone(&self.client);
         let digest = state.digest.clone();
         let max_size = self.get_max_batch_size();
@@ -237,29 +248,33 @@ impl RemoteService {
             .write()
             .await
             .push(tokio::spawn(async move {
-                if !blobs.is_empty() {
-                    if let Some(metadata) = &mut result.execution_metadata {
-                        metadata.output_upload_start_timestamp =
-                            create_timestamp(SystemTime::now());
+                if let Some(metadata) = &mut result.execution_metadata {
+                    metadata.output_upload_start_timestamp = create_timestamp(SystemTime::now());
+                }
+
+                // Don't save the action result if some of the blobs failed to upload
+                match batch_upload_blobs(client.clone(), digest.clone(), blobs, max_size as usize)
+                    .await
+                {
+                    Ok(uploaded) => {
+                        if !uploaded {
+                            return;
+                        }
                     }
+                    Err(error) => {
+                        warn!(
+                            hash = &digest.hash,
+                            "Failed to upload blobs and cache action result: {}",
+                            color::muted_light(error.to_string()),
+                        );
 
-                    let upload_result = batch_upload_blobs(
-                        client.clone(),
-                        digest.clone(),
-                        blobs,
-                        max_size as usize,
-                    )
-                    .await;
-
-                    // Don't save the action result if some of the blobs failed to upload
-                    if upload_result.is_err() || upload_result.is_ok_and(|res| !res) {
                         return;
                     }
+                };
 
-                    if let Some(metadata) = &mut result.execution_metadata {
-                        metadata.output_upload_completed_timestamp =
-                            create_timestamp(SystemTime::now());
-                    }
+                if let Some(metadata) = &mut result.execution_metadata {
+                    metadata.output_upload_completed_timestamp =
+                        create_timestamp(SystemTime::now());
                 }
 
                 if let Err(error) = client.update_action_result(&digest, result).await {
@@ -271,27 +286,44 @@ impl RemoteService {
                 }
             }));
 
-        Ok(())
+        // We don't actually know at this point if they all uploaded
+        Ok(true)
     }
 
     #[instrument(skip(self, state))]
-    pub async fn restore_action_result(&self, state: &mut ActionState<'_>) -> miette::Result<()> {
+    pub async fn restore_action_result(&self, state: &mut ActionState<'_>) -> miette::Result<bool> {
         if !self.cache_enabled {
-            return Ok(());
+            return Ok(false);
         }
 
         let Some(result) = &mut state.action_result else {
-            return Ok(());
+            return Ok(false);
         };
 
-        batch_download_blobs(
+        match batch_download_blobs(
             self.client.clone(),
             &state.digest,
             result,
             &self.workspace_root,
             self.get_max_batch_size() as usize,
         )
-        .await?;
+        .await
+        {
+            Ok(downloaded) => {
+                if !downloaded {
+                    return Ok(false);
+                }
+            }
+            Err(error) => {
+                warn!(
+                    hash = &state.digest.hash,
+                    "Failed to download blobs and restore action result: {}",
+                    color::muted_light(error.to_string()),
+                );
+
+                return Ok(false);
+            }
+        };
 
         // The stderr/stdout blobs may not have been inlined,
         // so we need to fetch them manually
@@ -315,6 +347,10 @@ impl RemoteService {
                 .batch_read_blobs(&state.digest, stdio_digests)
                 .await?
             {
+                let Some(blob) = blob else {
+                    continue;
+                };
+
                 if result
                     .stderr_digest
                     .as_ref()
@@ -334,7 +370,7 @@ impl RemoteService {
             }
         }
 
-        Ok(())
+        Ok(true)
     }
 
     #[instrument(skip(self))]
@@ -383,26 +419,10 @@ async fn batch_upload_blobs(
         // Streaming
         if group.stream {
             set.spawn(async move {
-                match client
+                client
                     .stream_update_blob(&action_digest, group.items.remove(0))
                     .await
-                {
-                    Ok(result) => {
-                        if result.is_some() {
-                            return true;
-                        }
-                    }
-                    Err(error) => {
-                        warn!(
-                            hash = &action_digest.hash,
-                            group = group_index + 1,
-                            "Failed to stream upload blob: {}",
-                            color::muted_light(error.to_string()),
-                        );
-                    }
-                };
-
-                false
+                    .map(|res| vec![Some(res)])
             });
 
             continue;
@@ -414,37 +434,26 @@ async fn batch_upload_blobs(
                 hash = &action_digest.hash,
                 blobs = group.items.len(),
                 size = group.size,
-                max_size,
                 "Batching blobs upload (group {} of {})",
                 group_index + 1,
                 group_total
             );
         }
 
-        set.spawn(async move {
-            match client.batch_update_blobs(&action_digest, group.items).await {
-                Ok(result) => {
-                    if result.into_iter().all(|res| res.is_some()) {
-                        return true;
-                    }
-                }
-                Err(error) => {
-                    warn!(
-                        hash = &action_digest.hash,
-                        group = group_index + 1,
-                        "Failed to upload blobs: {}",
-                        color::muted_light(error.to_string()),
-                    );
-                }
-            };
-
-            false
-        });
+        set.spawn(async move { client.batch_update_blobs(&action_digest, group.items).await });
     }
 
-    let results = set.join_all().await;
+    while let Some(res) = set.join_next().await {
+        for maybe_digest in res.into_diagnostic()?? {
+            if maybe_digest.is_none() {
+                set.abort_all();
 
-    Ok(results.into_iter().all(|passed| passed))
+                return Ok(false);
+            }
+        }
+    }
+
+    Ok(true)
 }
 
 async fn batch_download_blobs(
@@ -453,7 +462,7 @@ async fn batch_download_blobs(
     result: &ActionResult,
     workspace_root: &Path,
     max_size: usize,
-) -> miette::Result<()> {
+) -> miette::Result<bool> {
     let mut file_map = FxHashMap::default();
     let mut blob_digests = vec![];
 
@@ -469,11 +478,11 @@ async fn batch_download_blobs(
         partition_into_groups(blob_digests, max_size, |dig| dig.size_bytes as usize);
 
     if digest_groups.is_empty() {
-        return Ok(());
+        return Ok(true);
     }
 
     let group_total = digest_groups.len();
-    let mut set = JoinSet::<miette::Result<Vec<Blob>>>::default();
+    let mut set = JoinSet::<miette::Result<Vec<Option<Blob>>>>::default();
 
     for (group_index, mut group) in digest_groups.into_iter() {
         let client = Arc::clone(&client);
@@ -482,16 +491,10 @@ async fn batch_download_blobs(
         // Streaming
         if group.stream {
             set.spawn(async move {
-                let mut blobs = vec![];
-
-                if let Some(blob) = client
+                client
                     .stream_read_blob(&action_digest, group.items.remove(0))
-                    .await?
-                {
-                    blobs.push(blob);
-                }
-
-                Ok(blobs)
+                    .await
+                    .map(|res| vec![res])
             });
 
             continue;
@@ -515,6 +518,12 @@ async fn batch_download_blobs(
 
     while let Some(res) = set.join_next().await {
         for blob in res.into_diagnostic()?? {
+            let Some(blob) = blob else {
+                set.abort_all();
+
+                return Ok(false);
+            };
+
             if let Some(file) = file_map.get(&blob.digest.hash) {
                 write_output_file(workspace_root.join(&file.path), blob.bytes, file)?;
             }
@@ -531,7 +540,7 @@ async fn batch_download_blobs(
         )?;
     }
 
-    Ok(())
+    Ok(true)
 }
 
 struct Partition<T> {
@@ -547,19 +556,32 @@ fn partition_into_groups<T>(
 ) -> BTreeMap<i32, Partition<T>> {
     let mut groups = BTreeMap::<i32, Partition<T>>::default();
 
+    // If the max size is larger than 2mb, we reduce the
+    // group overall size by half, so that we divide blobs
+    // across multiple groups, allowing them to be parallelized
+    // better. Waiting for a 2mb up/download is much slower
+    // than waiting for multiple parallel 500kb up/downloads.
+    let max_group_size = if max_size >= 4194304 {
+        2097144
+    } else if max_size > 2097144 {
+        max_size / 2
+    } else {
+        max_size
+    };
+
     for item in items {
         let item_size = get_size(&item);
         let mut index_to_use = -1;
         let mut stream = false;
 
         // Item is too large, must be streamed
-        if item_size >= max_size {
+        if item_size >= max_group_size {
             stream = true;
         }
         // Try and find a partition that this item can go into
         else {
             for (index, group) in &groups {
-                if !group.stream && (group.size + item_size) <= max_size {
+                if !group.stream && (group.size + item_size) <= max_group_size {
                     index_to_use = *index;
                     break;
                 }